# PERF/CLN: restrict _isna_ndarraylike to actual ndarray/EA