xref https://github.com/pandas-dev/pandas/issues/39146

I am looking at some `isna` related benchmarks, and noticing a lot of overhead related to checking of types. This makes `_isna_ndarraylike` stricter to only handle ndarray and EA, and not Series/Index (and therefore also renamed it to `_isna_array`).

Reproducer from the `frame_methods.py::Isnull` benchmark:

```python
N = 10 ** 3
sample = np.array([np.nan, 1.0])
data = np.random.choice(sample, (N, N))
df = DataFrame(data)
df_am = df._as_manager("array")
```

```
In [3]: %timeit df_am.isna()
5.03 ms ± 125 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)  <-- master
4.25 ms ± 9.94 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)  <-- PR
```

This gives a small improvement. But it further allows to directly use `_is_array` instead of the general `isna` from Manager.isna, which will give further improvements (but was planning to do this as a follow-up)